게시물 검색키워드타입 추가

+++ src/main/java/com/project/untact/controller/UsrArticleController.java
@@ -1,90 +1,103 @@
 package com.project.untact.controller;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.project.untact.dto.Article;
 import com.project.untact.dto.ResultData;
-import com.project.untact.service.ArticleService;
 import com.project.untact.util.Util;
+import com.project.untact.service.ArticleService;
 
 @Controller
 public class UsrArticleController {
-	
+	@Autowired
 	private ArticleService articleService;
-	
+
 	@RequestMapping("/usr/article/detail")
 	@ResponseBody
 	public Article showDetail(int id) {
 		Article article = articleService.getArticle(id);
-		
+
 		return article;
 	}
-	
+
 	@RequestMapping("/usr/article/list")
 	@ResponseBody
-	public List<Article> showList(String searchKeyword) {
+	public List<Article> showList(String searchKeywordType, String searchKeyword) {
+		if ( searchKeywordType != null ) {
+			searchKeywordType = searchKeywordType.trim();
+		}
 		
-		if( searchKeyword != null && searchKeyword.length() == 0) {
+		if ( searchKeywordType == null || searchKeywordType.length() == 0 ) {
+			searchKeywordType = "titleAndBody";
+		}
+		
+		if ( searchKeyword != null && searchKeyword.length() == 0 ) {
 			searchKeyword = null;
 		}
 		
-		if( searchKeyword != null ) {
+		if ( searchKeyword != null ) {
 			searchKeyword = searchKeyword.trim();
 		}
 		
-		return articleService.getArticles(searchKeyword);
+		return articleService.getArticles(searchKeywordType, searchKeyword);
 	}
-	
+
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
 	public ResultData doAdd(String title, String body) {
 		if (title == null) {
 			return new ResultData("F-1", "title을 입력해주세요.");
 		}
+
 		if (body == null) {
 			return new ResultData("F-1", "body를 입력해주세요.");
 		}
+
 		return articleService.add(title, body);
 	}
-	
+
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
 	public ResultData doDelete(Integer id) {
+		if (id == null) {
+			return new ResultData("F-1", "id를 입력해주세요.");
+		}
+
 		Article article = articleService.getArticle(id);
-		
+
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
-		
+
 		return articleService.deleteArticle(id);
 	}
 
-	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
 	public ResultData doModify(Integer id, String title, String body) {
 		if (id == null) {
-			return new ResultData("F-1", "id을 입력해주세요.");
+			return new ResultData("F-1", "id를 입력해주세요.");
 		}
+
 		if (title == null) {
 			return new ResultData("F-1", "title을 입력해주세요.");
 		}
+
 		if (body == null) {
 			return new ResultData("F-1", "body를 입력해주세요.");
 		}
-		
+
 		Article article = articleService.getArticle(id);
+
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
-		return articleService.modify(id, title, body);
 
+		return articleService.modify(id, title, body);
 	}
 }
